dsa-indexer(npu-move,mechanical_provable): split NPU indexer logic
